Restructuring lua scripts
