+++ lua/modules/LeviatanScans.lua
@@ -1,0 +1,73 @@
+----------------------------------------------------------------------------------------------------
+-- Scripting Parameters
+----------------------------------------------------------------------------------------------------
+
+local LuaDebug   = require 'Modules.LuaDebugging'
+-- LuaDebugging  = true   --> Override the global LuaDebugging variable by uncommenting this line.
+-- LuaStatistics = true   --> Override the global LuaStatistics variable by uncommenting this line.
+
+
+----------------------------------------------------------------------------------------------------
+-- Local Constants
+----------------------------------------------------------------------------------------------------
+
+local Template   = require 'Modules.Template-Genkan'
+-- DirectoryPagination = '/'   --> Override template variable by uncommenting this line.
+
+
+----------------------------------------------------------------------------------------------------
+-- Event Functions
+----------------------------------------------------------------------------------------------------
+
+-- Get info and chapter list for current manga.
+function GetInfo()
+  Template.GetInfo()
+  
+  return no_error
+end
+
+
+-- Get the page count of the manga list of the current website.
+function GetDirectoryPageNumber()
+  Template.GetDirectoryPageNumber()
+  
+  return no_error
+end
+
+
+-- Get links and names from the manga list of the current website.
+function GetNameAndLink()
+  Template.GetNameAndLink()
+  
+  return no_error
+end
+
+
+-- Get the page count for the current chapter.
+function GetPageNumber()
+  Template.GetPageNumber()
+  
+  return no_error
+end
+
+
+----------------------------------------------------------------------------------------------------
+-- Module Initialization
+----------------------------------------------------------------------------------------------------
+
+function Init()
+  AddWebsiteModule('LeviatanScans',   'https://es.leviatanscans.com', 'Spanish-Scanlation')
+  AddWebsiteModule('LeviatanScansEN', 'https://leviatanscans.com',    'English-Scanlation')
+end
+
+function AddWebsiteModule(name, url, category)
+  local m = NewModule()
+  m.Website                  = name
+  m.RootURL                  = url
+  m.Category                 = category
+  m.OnGetInfo                = 'GetInfo'
+  m.OnGetDirectoryPageNumber = 'GetDirectoryPageNumber'
+  m.OnGetNameAndLink         = 'GetNameAndLink'
+  m.OnGetPageNumber          = 'GetPageNumber'
+  return m
+end
